Developing an ellipsoid model of the Dobot and trying to calculate intersecting points with cube
Developing an ellipsoid model of the Dobot and trying to calculate intersecting points with cube. Need to work out how to calculate the geometric distance and apply the angle of the ellipsoid to the calculation. At present, I think it is calculating the geometric distance using global coordinates, not local