Implemented ellipsoid collision checking for each step of the calculated trajectory
Implemented ellipsoid collision checking by iterating through each step of the calculated trajectory and checking if any ellipsoids collide with the cubes in the environment. Works with multiple cubes in the environment. Could be handy if we can work out how to implement a cube point cloud around every object in the environment and then do a simple check each time, iterating through each cube and making sure there are no collisions.